requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
